+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Changelog
 
-## 0.3.3
+## 0.4.0
 - Changed `Pipeline.__bool__()` to return whether the pipeline contains configured steps instead of whether it is currently running.
 - Added `Stack.put()` as a compatibility for `Stack.push()`.
 - Added `Pipeline.reverse()` for in-place reversal of pipeline steps.
 - Added `reversed()` support for iterating over pipeline steps in reverse order.
 - Added `Pipeline.update()` for adding multiple pipeline steps at once with full validation before insertion.
-- Added support  appending a single step with the `+=` operator.
+- Added support for appending a single step with the `+=` operator.
 - Added support creating a new pipeline by concatenating steps with the `+` operator.
 - Added support prepending pipeline steps with the `+` operator.
 - Added support comparing pipeline configurations with the `==` operator.